+++ out/extension.js
@@ -33,30 +33,42 @@ const openSection_1 = require("./commands/openSection");
 const codeLensProvider_1 = require("./providers/codeLensProvider");
 const addIconClass_1 = require("./commands/addIconClass");
 const convertToNetlifyForm_1 = require("./commands/convertToNetlifyForm");
+const optimizeSharpImages_1 = require("./commands/optimizeSharpImages");
 function activate(context) {
-    console.log('codestitch-helper is now active!');
+    console.log("codestitch-helper is now active!");
     const sectionNavProvider = new sectionNavigationProvider_1.SectionNavigationProvider();
-    vscode.window.registerTreeDataProvider('codestitchHelper', sectionNavProvider);
-    const openSectionDisposable = vscode.commands.registerCommand('codestitchHelper.openSection', (section) => {
+    vscode.window.registerTreeDataProvider("codestitchHelper", sectionNavProvider);
+    const openSectionDisposable = vscode.commands.registerCommand("codestitchHelper.openSection", (section) => {
         (0, openSection_1.openSection)(section);
     });
-    const selectAllDisposable = vscode.commands.registerCommand('codestitchHelper.selectAll', (section) => {
+    const selectAllDisposable = vscode.commands.registerCommand("codestitchHelper.selectAll", (section) => {
         (0, selectAll_1.selectAll)(section);
     });
-    const replaceNavTabsDisposable = vscode.commands.registerCommand('codestitchHelper.replaceNavTabs', (document) => {
+    const replaceNavTabsDisposable = vscode.commands.registerCommand("codestitchHelper.replaceNavTabs", (document) => {
         (0, replaceNavTabs_1.replaceNavTabs)(document);
     });
-    const addIconClassDisposable = vscode.commands.registerCommand('codestitchHelper.addIconClass', (document, range) => {
+    const addIconClassDisposable = vscode.commands.registerCommand("codestitchHelper.addIconClass", (document, range) => {
         (0, addIconClass_1.addIconClass)(document, range);
     });
-    const convertToNetlifyFormDisposable = vscode.commands.registerCommand('codestitchHelper.convertToNetlifyForm', (document) => {
+    const convertToNetlifyFormDisposable = vscode.commands.registerCommand("codestitchHelper.convertToNetlifyForm", (document) => {
         (0, convertToNetlifyForm_1.convertToNetlifyForm)(document);
     });
+    const optimizeSharpImagesDisposable = vscode.commands.registerCommand("codestitchHelper.optimizeSharpImages", (document, range) => {
+        (0, optimizeSharpImages_1.optimizeSharpImages)(document, range);
+    });
     // Ensure the Explorer view is open
-    vscode.commands.executeCommand('workbench.view.explorer'); // No need to focus on a specific view
+    vscode.commands.executeCommand("workbench.view.explorer"); // No need to focus on a specific view
     // Register CodeLensProvider for both HTML and SCSS
-    context.subscriptions.push(vscode.languages.registerCodeLensProvider({ language: '*' }, new codeLensProvider_1.CodeLensProvider()));
-    const supportedLanguages = ['scss', 'css', 'sass', 'less', 'njk', 'nunjucks', 'html'];
+    context.subscriptions.push(vscode.languages.registerCodeLensProvider({ language: "*" }, new codeLensProvider_1.CodeLensProvider()));
+    const supportedLanguages = [
+        "scss",
+        "css",
+        "sass",
+        "less",
+        "njk",
+        "nunjucks",
+        "html",
+    ];
     vscode.workspace.onDidChangeTextDocument((event) => {
         if (supportedLanguages.includes(event.document.languageId)) {
             sectionNavProvider.refresh();
@@ -72,7 +84,7 @@ function activate(context) {
             sectionNavProvider.refresh();
         }
     });
-    context.subscriptions.push(openSectionDisposable, selectAllDisposable, replaceNavTabsDisposable, addIconClassDisposable, convertToNetlifyFormDisposable);
+    context.subscriptions.push(optimizeSharpImagesDisposable, openSectionDisposable, selectAllDisposable, replaceNavTabsDisposable, addIconClassDisposable, convertToNetlifyFormDisposable);
 }
 function deactivate() { }
 //# sourceMappingURL=extension.js.map